+++ CHANGELOG.md
@@ -2,11 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [0.8.6]
+### Changed
+- serialize `item.properties` into a string when saving items
 ### Added
 - sharing controls on item detail view in dummy
 - added `surge` and `github` environment settings
 - updated `npm run deploy` to use the `github` environment
+- additional group controls in dummy
 
 ## [0.8.5]
 ### Added